Add configuration setting for reflowing doc comments
This change reflows documentation comments to the desired line length
when the setting is enabled.

swift-markdown isn't ideal for this purpose, since it's reformatter
doesn't really maintain fidelity to the original (quotes are smartified,
double-backticks get switched to single, etc.), but this hand-rolled
Markdown classifier probably isn't tenable for actual adoption.